+++ src/lib.rs
@@ -125,6 +125,7 @@ pub mod outline_panel;
 pub mod pomodoro_tool;
 pub mod right_dock;
 pub mod roam;
+pub mod workspace;
 pub mod spellcheck;
 pub mod status_bar_panel;
 pub mod system_information_panel;